% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Countries.R
\docType{data}
\name{Countries}
\alias{Countries}
\title{Countries}
\format{
A dataframe with 250 rows and 6 columns.
}
\usage{
Countries
}
\description{
Countries data with country code, name, currency code,
  population, capital and continent name.
}
\keyword{datasets}
