% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jstree.R
\name{jstree}
\alias{jstree}
\title{<Add Title>}
\usage{
jstree(
  data,
  width = NULL,
  height = NULL,
  elementId = NULL,
  checkbox = FALSE,
  search = FALSE,
  searchtime = 250,
  dragAndDrop = FALSE,
  dnd = NULL,
  types = NULL,
  sort = FALSE,
  unique = FALSE,
  wholerow = FALSE,
  contextMenu = FALSE,
  checkCallback = NULL
)
}
\description{
<Add Description>
}
