% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jstree-shiny.R
\name{jstreeDestroy}
\alias{jstreeDestroy}
\title{Destroy jstree}
\usage{
jstreeDestroy(session, id)
}
\arguments{
\item{session}{the Shiny \code{session} object}

\item{id}{the id of the tree to be destroyed}
}
\value{
No value, just called to destroy a tree.
}
\description{
Destroy a `jstree` instance.
}
