% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jstreeExamples.R
\name{jstreeExample}
\alias{jstreeExample}
\title{Run a Shiny jsTreeR example}
\usage{
jstreeExample(example, display.mode = "showcase", ...)
}
\arguments{
\item{example}{example name}

\item{display.mode}{the display mode to use when running the example; see
\code{\link[shiny:runApp]{runApp}}}

\item{...}{arguments passed to \code{\link[shiny:runApp]{runApp}}}
}
\value{
No return value, just launches a Shiny app.
}
\description{
A function to run examples of Shiny apps using the
  \code{jsTreeR} package.
}
\examples{
if(interactive()){
  jstreeExample("folder")
}
if(interactive()){
  jstreeExample("fontawesome")
}
if(interactive()){
  jstreeExample("SuperTinyIcons")
}
if(interactive()){
  jstreeExample("filtering")
}
if(interactive()){
  jstreeExample("grid")
}
if(interactive()){
  jstreeExample("gridFiltering")
}
}
