% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jstreeExamples.R
\name{jstreeExamples}
\alias{jstreeExamples}
\title{jsTreeR examples}
\usage{
jstreeExamples()
}
\value{
No returned value, just prints a message listing the example names.
}
\description{
List of Shiny examples.
}
\examples{
jstreeExamples()
if(interactive()){
  jstreeExample("grid")
}
}
