% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jstree-shiny.R
\name{jstreeUpdate}
\alias{jstreeUpdate}
\title{Update jstree}
\usage{
jstreeUpdate(session, id, nodes)
}
\arguments{
\item{session}{the Shiny \code{session} object}

\item{id}{the id of the tree to be updated}

\item{nodes}{the new \code{nodes} list}
}
\value{
No value, just called to update a tree.
}
\description{
Update a `jstree` instance in a Shiny app.
}
