% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeNavigator.R
\name{treeNavigator}
\alias{treeNavigator}
\alias{treeNavigatorUI}
\alias{treeNavigatorServer}
\title{Tree navigator (Shiny module)}
\usage{
treeNavigatorUI(id, width = "100\%", height = "auto")

treeNavigatorServer(
  id,
  rootFolder,
  search = TRUE,
  pattern = NULL,
  all.files = TRUE
)
}
\arguments{
\item{id}{xxx}

\item{width}{xxx}

\item{height}{xxx}

\item{rootFolder}{xx}

\item{search}{xx}

\item{pattern}{xx}

\item{all.files}{xx}
}
\value{
xxx
}
\description{
xxx
}
\examples{
###
}
