% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folderGadget.R
\name{folderGadget}
\alias{folderGadget}
\title{Folder gadget}
\usage{
folderGadget(dirs = ".", tabs = FALSE, recursive = TRUE, all.files = FALSE)
}
\arguments{
\item{dirs}{character vector of paths to some folders}

\item{tabs}{logical, whether to display the trees in tabs; this option is
effective only when there are two folders in the \code{dirs} argument}

\item{recursive, all.files}{options passed to \code{\link{list.files}};
even if \code{all.files = TRUE}, \code{'.git'} and \code{'.Rproj.user'}
folders are always discarded}
}
\description{
Shiny gadget allowing to manipulate one or more folders.
}
\note{
You can run the gadget for the current directory from the Addins menu
  within RStudio ('Explore current folder').
}
